backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
cx_Oracle==8.3.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
plotly==5.18.0
dash==2.14.2
gunicorn==21.2.0
SQLAlchemy==2.0.25
marshmallow==3.20.1 